Moved buildEnvironmentFromScript, check for vcast license, and setCodedTestOption to the vcastAdapter
